third_party/blink/web_tests/fast/multicol/vertical-lr/client-rect-after-spanner-expected.txt

Test that getBoundingClientRect works as expected on content after a spanner, writing-mode is vertical-lr

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementById('elm1').getBoundingClientRect().left is 200
PASS document.getElementById('elm1').getBoundingClientRect().top is 100
PASS document.getElementById('elm2').getBoundingClientRect().left is 250
PASS document.getElementById('elm2').getBoundingClientRect().top is 100
PASS document.getElementById('elm3').getBoundingClientRect().left is 200
PASS document.getElementById('elm3').getBoundingClientRect().top is 216
PASS document.getElementById('elm4').getBoundingClientRect().left is 250
PASS document.getElementById('elm4').getBoundingClientRect().top is 216
PASS document.getElementById('elm5').getBoundingClientRect().left is 200
PASS document.getElementById('elm5').getBoundingClientRect().top is 332
PASS successfullyParsed is true

TEST COMPLETE
